introduce-score-request-handler(move): Hand scoring over to ScoreRequestHandler
Pure physical move per MECH_COMMIT_SPLIT. Cut the TokenizerManagerScoreMixin
class body (the scoring staticmethods already retyped to
self: "ScoreRequestHandler" in prep); paste into ScoreRequestHandler,
dropping the @staticmethod decorators and restoring plain self. Drop the
mixin base from TokenizerManager bases and delete the mixin file.

Caller prefix replacement:
``TokenizerManagerScoreMixin.<method>(self.tokenizer_manager.score_request_handler, ...)``
→ ``self.tokenizer_manager.score_request_handler.<method>(...)``.
Drop the now-unused TokenizerManagerScoreMixin import from the
entrypoint caller files.

Refactor chain ID: introduce-score-request-handler-move
PR-Title: Extract ScoreRequestHandler from TokenizerManager
