+++ python/sglang/srt/entrypoints/engine_score_mixin.py
@@ -14,8 +14,8 @@
 """
 Engine mixin that exposes score() and async_score() on the Engine class.
 
-These methods delegate to TokenizerManager.score_request() which is provided
-by TokenizerManagerScoreMixin.
+These methods delegate to ``self.tokenizer_manager.score_request_handler.score_request()``
+(provided by ScoreRequestHandler).
 """
 
 from typing import List, Optional, Union
@@ -24,9 +24,6 @@ import torch
 
 from sglang.srt.managers.tokenizer_manager_components.score_request_handler import (
     ScoreResult,
-)
-from sglang.srt.managers.tokenizer_manager_score_mixin import (
-    TokenizerManagerScoreMixin,
 )
 
 
@@ -75,8 +72,7 @@ class EngineScoreMixin:
             optional pooled_hidden_states tensors.
         """
         return self.loop.run_until_complete(
-            TokenizerManagerScoreMixin.score_request(
-                self.tokenizer_manager.score_request_handler,
+            self.tokenizer_manager.score_request_handler.score_request(
                 query=query,
                 items=items,
                 label_token_ids=label_token_ids,
@@ -103,8 +99,7 @@ class EngineScoreMixin:
         return_pooled_hidden_states: bool = False,
     ) -> ScoreResult:
         """Asynchronous version of score(). See score() for full documentation."""
-        return await TokenizerManagerScoreMixin.score_request(
-            self.tokenizer_manager.score_request_handler,
+        return await self.tokenizer_manager.score_request_handler.score_request(
             query=query,
             items=items,
             label_token_ids=label_token_ids,
