+++ python/sglang/srt/entrypoints/openai/serving_score.py
@@ -12,9 +12,6 @@ from sglang.srt.entrypoints.openai.protocol import (
     UsageInfo,
 )
 from sglang.srt.entrypoints.openai.serving_base import OpenAIServingBase
-from sglang.srt.managers.tokenizer_manager_score_mixin import (
-    TokenizerManagerScoreMixin,
-)
 
 logger = logging.getLogger(__name__)
 
@@ -70,8 +67,7 @@ class OpenAIServingScore(OpenAIServingBase):
                 else None
             )
 
-            result = await TokenizerManagerScoreMixin.score_request(
-                self.tokenizer_manager.score_request_handler,
+            result = await self.tokenizer_manager.score_request_handler.score_request(
                 query=request.query,
                 items=request.items,
                 label_token_ids=request.label_token_ids,
